requirements.txt
pydicom==2.2.2
opencv-python==4.5.5.62
Pillow==9.0.0
Flask==2.0.2
matplotlib==3.5.1
highdicom==0.14.0
numpy>=1.21
python-gdcm